refactor: Extract artifacts modules
