+++ src/artifacts/index.ts
@@ -1,0 +1,2 @@
+export * from './filter-artifacts';
+export * from './validate-webpack-stats';
