+++ src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from './debug';
-export * from './filter-artifacts';
 export * from './get-commit-message';
 export * from './get-env-vars';
 export * from './get-slug';
